user created by postman and bugs fixed

+++ src/utils/cloudinary.js
@@ -7,7 +7,6 @@ cloudinary.config({
         api_secret: process.env.CLOUDINARY_API_SECRET
 });
 
-
 const uploadOnCloudinary = async (localFilePath) => {
     try {
         if(!localFilePath) return null;
@@ -15,14 +14,16 @@ const uploadOnCloudinary = async (localFilePath) => {
         const response = await cloudinary.uploader.upload(localFilePath, {
             resource_type: "auto"
         });
-
         //file has been uploaded successfully
         console.log("file is uploaded on cloudinary ", response.url);
+        fs.unlinkSync(localFilePath);
+
         return response;
         
     } catch (error) {
-        fs.unlinkSync(localFilePath) //remove the locally saved temporary file as the upload operation failed
-        return null
+        console.log("Cloudinary upload error: ", error);
+        if(fs.existsSync(localFilePath)) fs.unlinkSync(localFilePath) //remove the locally saved temporary file as the upload operation failed
+        return null; 
     }
 }
